requirements.txt
datasets==2.15.0
huggingface_hub==0.19.4
hydra-core==1.3.2
numpy==1.26.2
tokenizers==0.15.0
transformers==4.36.0
omegaconf==2.3.0
openai==1.3.9
nltk==3.8.1
scikit-learn==1.3.2
gitpython==3.1.40
matplotlib==3.8.2
pandas~=2.2.0
wandb~=0.16.2
pytest~=8.0.1
langdetect~=1.0.9
tenacity~=8.2.3
python-dotenv~=1.0.1
anthropic~=0.21.3
tiktoken~=0.6.0
unidiff~=0.7.5
torch~=2.2.2